// src/app/helpers/enums.ts
export enum BoxEnum {
  OT = 1,
  CajillaDeSeguridad = 2,
}

export enum ElementClassificationEnum {
  Armas = 1,
  Capacitacion = 2,
  Comunicacion = 3,
  ExamenesHSE = 4,
  GastosContratacion = 5,
  DotacionPuesto = 6,
  SostenimientoCampo = 7,
  DotacionUniforme = 8,
  Vehiculos = 9,
  Otro = 10,
  ArmasElementos = 11,
  VehiculosElementos = 12,
  BonosOT = 13,
}

export enum Reporte {
  OfertaEconomica = 1,
  SalariosCotizacionesActivas = 2,
  ElementosCotizacionesActivas = 3,
  TrazabilidadCotizaciones = 4,
}